Fix probles up to 14

+++ src/problem009.cpp
@@ -5,16 +5,16 @@
 int main() {
     std::cout << "Solving Project Euler Problem 009" << std::endl;
     
-    long long n = 1000;
-    for (triple<long long> pt: pythagorean_triples(1000)) {
-        long long s = pt.a + pt.b + pt.c;
+    ull n = 1000;
+    for (triple<ull> pt: pythagorean_triples(1000)) {
+        ull s = pt.a + pt.b + pt.c;
         if (n % s != 0)
             continue;
-        long long k = n / s;
-        long long a = k * pt.a;
-        long long b = k * pt.b;
-        long long c = k * pt.c;
-        long long p = a * b * c;
+        ull k = n / s;
+        ull a = k * pt.a;
+        ull b = k * pt.b;
+        ull c = k * pt.c;
+        ull p = a * b * c;
         std::cout << p << std::endl;
     }
